+++ app/src/main/kotlin/com/firstpenguin/app/domain/emotion/dto/TagResponse.kt
@@ -1,4 +1,4 @@
-package com.firstpenguin.app.emotion.dto
+package com.firstpenguin.app.domain.emotion.dto
 
 data class TagResponse(
     val tagList: List<TagDto>
